+++ src/main/java/minecrafttransportsimulator/blocks/components/IBlockTileEntity.java
@@ -1,6 +1,6 @@
 package minecrafttransportsimulator.blocks.components;
 
-import minecrafttransportsimulator.baseclasses.Point3i;
+import minecrafttransportsimulator.baseclasses.Point3d;
 import minecrafttransportsimulator.blocks.tileentities.components.ATileEntityBase;
 import minecrafttransportsimulator.jsondefs.AJSONItem;
 import minecrafttransportsimulator.mcinterface.WrapperNBT;
@@ -16,12 +16,12 @@ public interface IBlockTileEntity<TileEntityType extends ATileEntityBase<? exten
 	/**
 	 *  Gets a new Tile Entity for this block.
 	 */
-	public TileEntityType createTileEntity(WrapperWorld world, Point3i position, WrapperNBT data);
+	public TileEntityType createTileEntity(WrapperWorld world, Point3d position, WrapperNBT data);
 	
 	/**
 	 *  Gets the class that this Tile Entity is made from.
 	 *  This is for registration, not construction.  For construction,
-	 *  use {@link #createTileEntity(WrapperWorld, Point3i, WrapperNBT)}
+	 *  use {@link #createTileEntity(WrapperWorld, Point3d, WrapperNBT)}
 	 */
 	public Class<TileEntityType> getTileEntityClass();
 }
